Add load query to query server
